+++ CHANGES.md
@@ -1,3 +1,5 @@
+- v0.13.1
+    - `unikmer count`: small performance improvement.
 - v0.13.0
     - new command `unikmer common`: Finding k-mers shared by most of multiple binary files.
     - `unikmer common/count/diff/grep/rfilter/sort/split/union`: faster sorting.